Classes reorganized so that TPC, SVT, SSD, FTPC classes are derived from a single set of interfaces; remianing detectors each have their own interface (defined by the detector group).
The following applies to EMC, RICH, TRG:
The xxx_Reader constructor is passed a pointer to its top-level pointer
bank (xxxP). All code which existed inside the detector-specific classes
which served to locate this top-level pointer bank has now been extracted and
moved to the EventReader::findBank() method.

Existing user code for the TPC, SVT, FTPC, SSD should not be changed by this
re-organization. Other detectors now have a new API. The API for each of
these "other" detectors is now completely free of constraints, with the
exception of the arguments passed to the constructor xxx_Reader::xxx_Reader()