Changes to deal with seg. faults in the file name handling: Conventions: StMiniMcMaker looks for the input file from the IO maker to figure out if the file has changed.  This is done using TString::Contains() in Make(). Usually we will run one file at a time, but in order not to break Bum's scheme of being able to process several files in one go, this is left as is.  However, for embedding, the file name is not enough, in Eric's new scheme there are repeated file names.  This is resolved by adding a prefix to the output file name.  However, this prefix should not be overwritten, so the current code only replaces the string inside the output file name pertaining to the input file name, and leaves the prefix of the output file intact.  This was done for embedding looking for st_physics, and here is where the problem arose: hijing files begin with a different prefix.  To solve this problem, the input file name prefix is now an input parameter in the macro.
StMiniEmbed.C and StMiniHijing.C now conform to this convention.  StMiniEmbed.C
did not change its prototype, because all embedding files have st_phyics as prefix.
StMiniHijing.C changed its prototype, now it takes as an input argument the prefix,
but in order not to break Jenn's scripts if she was already using this macro,
this parameter was added at the end and defaults to "rcf", which is appropriate
for hijing files reconstructed in rcf.