Added CSMStatusUtils to the StEmcPool directory.  -David Relyea (DRR)
CSMStatusUtils contains StBemcStatusMaker, which runs on MuDsts and puts
data from all calorimeter channels into 2d histograms, which are then read
out by CSMBuildRunMap and CSMStatusUtils.  The latter then does QA work to
determine whether or not calorimeter channels are good/bad.  Currently, it
spits out data (status files, hot tower plots, html files) into rather
haphazard places - I need to fix this so that the directory structure is
checked before it tries to write data.  It also assigns too many "bad"
flags for channels which are likely fine, but which happen not to have
data in certain bits.  I'm working on it.