Needed to reorganize the code in order to make it VMC compatible. The old structure had an incosistent path to hits, and VMC had trouble accessing the hits data.
The 3 sectors on the PMD are now described by the same block, and there
is logic to ensure they are inflated correctly according to a flag (so we
create proper geometry just like before. There is no distinct  PHMT, it has
been absorbed into PHMS. In addition, formatting improvements have been made.