Implemented $FUZZ4C, a fuzz factor allow to boost up the disk usage for C (calibration) pass. Well, it got me that far (but now data09 is 100% full and no calib is possible) ...
Last year logique for PanFS was not entirely correct i.e. a df would
be done anyway which is a waste of time. Added explaination.

Miscellaneous message reshape: no big deal but the report option of
the FastOffline browser showed too long lines or lines that meant
nothing to me (even).