Create a copy of sensor transformation matrix using the ideal TGeo geometry
We need to create a copy of the transformation matrix in order to have an
ability to modify it without disturbing the original TGeo geometry. The
modifications as in the next commit are needed to manually adjust the placement
of the sensor layers. This change introduces no new functionality and is done
only to introduce the subsequent change.