Convert assert to warning: |mJoinPars.hz() - mTargetHz| <= 1e-10
The assert was checking the magnetic field in the updated Sti node during the
refit. With the addition of iTPC layers some CA provided tracks do not look
healthy with nodes coming from unrelated sectors. This is not fully understood
yet but believed to be the reason failing the assert